requirements.txt
fastapi
uvicorn
openai-whisper
pymupdf
pytesseract
Pillow
python-multipart
python-dotenv
torch
setuptools
